SOLR-9506 Parallalize/Cache per segment fingerprint
